+++ HellfireControl/src/HellfireControl/Core/Common.hpp
@@ -9,7 +9,7 @@
 #include <thread>
 
 //Borrowed from Vulkan docs
-#define HC_CONVERT_TO_VERSION_NO(_variant, _major, _minor, _patch)  ((((uint32_t)(_variant)) << 29U) | (((uint32_t)(_major)) << 22U) | (((uint32_t)(_minor)) << 12U) | ((uint32_t)(_patch)))
+#define HC_CONVERT_TO_VERSION_NO(_variant, _major, _minor, _patch)  ((((uint64_t)(_variant)) << 29U) | (((uint64_t)(_major)) << 22U) | (((uint64_t)(_minor)) << 12U) | ((uint64_t)(_patch)))
 //Engine Version No.
 #define HC_ENGINE_VERSION HC_CONVERT_TO_VERSION_NO(1, 1, 0, 0)
